+++ src/index.html
@@ -42,7 +42,7 @@
     <div class="container">
       <main class="main__page">
         <section class="section_main">
-          <ul class="gallery js-gallery"></ul>
+          <ul class="gallery list js-gallery"></ul>
           <div class="pagination-thumb">
             <div id="tui-pagination-container" class="tui-pagination"></div>
           </div>
