+++ src/sass/base/_container.scss
@@ -1,7 +1,7 @@
 .container {
   // Контейнер для mobil
   margin: 0 auto;
-  padding: 0 20px;
+  padding: 60px 20px;
   width: 320px;
 
   // Контейнер для tablet
