+++ src/sass/main.scss
@@ -11,6 +11,7 @@
 @import './layout/header';
 @import './base/footer';
 @import './componets/pagination';
+@import './layout/section_main';
 
 // Секции и блоки -> layouts
