+++ src/templates/cards.hbs
@@ -1,7 +1,7 @@
 {{#each this}}
-<li>
+<li class="gallery__item">
     <div class="movie-card">
-        <img src="https://image.tmdb.org/t/p/w300{{poster_path}}" alt="" />
+        <img class="poster" src="https://image.tmdb.org/t/p/w300{{poster_path}}" alt="" />
         <div class="movie-details">
             <p class="movie-name">{{original_title}}</p>
             <p class="movie-info">{{genre_ids}} | {{release_date}}</p>
